+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/runtime/ServiceRuntimeIntegrationTest.java
@@ -19,6 +19,7 @@ package com.exonum.binding.core.runtime;
 import static com.exonum.binding.core.runtime.ServiceWrapper.DEFAULT_INTERFACE_NAME;
 import static com.exonum.binding.test.Bytes.bytes;
 import static com.google.common.collect.Comparators.isInStrictOrder;
+import static java.util.Collections.emptyList;
 import static java.util.Comparator.comparing;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -42,6 +43,8 @@ import com.exonum.binding.core.proxy.CloseFailuresException;
 import com.exonum.binding.core.service.BlockCommittedEvent;
 import com.exonum.binding.core.service.Configuration;
 import com.exonum.binding.core.service.ExecutionContext;
+import com.exonum.binding.core.service.migration.AbstractMigrationScript;
+import com.exonum.binding.core.service.migration.MigrationScript;
 import com.exonum.binding.core.storage.database.Database;
 import com.exonum.binding.core.storage.database.Fork;
 import com.exonum.binding.core.storage.database.Snapshot;
@@ -52,7 +55,6 @@ import com.exonum.messages.core.runtime.Lifecycle.InstanceStatus.Simple;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -63,6 +65,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentCaptor;
@@ -74,9 +77,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 @ExtendWith(MockitoExtension.class)
 class ServiceRuntimeIntegrationTest {
 
-  // [ECR-587] Replace with a temp directory obtained from a TempDir JUnit extension so that
-  //   the check of its existence passes.
-  static final Path ARTIFACTS_DIR = Paths.get("/tmp/");
+  @TempDir
+  static Path ARTIFACTS_DIR;
   static final String TEST_NAME = "test_service_name";
   static final int TEST_ID = 17;
   static final HashCode TEST_HASH = HashCode.fromBytes(bytes(1, 2, 3));
@@ -116,7 +118,7 @@ class ServiceRuntimeIntegrationTest {
     String artifactFilename = "foo-service.jar";
     Path serviceArtifactLocation = ARTIFACTS_DIR.resolve(artifactFilename);
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(serviceId, TestServiceModule::new);
+        .newInstance(serviceId, TestServiceModule::new, emptyList());
     when(serviceLoader.loadService(serviceArtifactLocation))
         .thenReturn(serviceDefinition);
     when(serviceLoader.findService(serviceId))
@@ -136,7 +138,7 @@ class ServiceRuntimeIntegrationTest {
     String artifactFilename = "foo-service.jar";
     Path serviceArtifactLocation = ARTIFACTS_DIR.resolve(artifactFilename);
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(actualId, TestServiceModule::new);
+        .newInstance(actualId, TestServiceModule::new, emptyList());
     when(serviceLoader.loadService(serviceArtifactLocation))
         .thenReturn(serviceDefinition);
 
@@ -176,7 +178,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.parseFrom("1:com.acme/foo-service:1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -234,7 +236,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.newJavaId("com.acme/foo-service", "1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -267,7 +269,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.newJavaId("com.acme/foo-service", "1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -301,7 +303,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.newJavaId("com.acme/foo-service", "1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -348,7 +350,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.newJavaId("com.acme/foo-service", "1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -408,7 +410,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.parseFrom("1:com.acme/foo-service:1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -444,7 +446,7 @@ class ServiceRuntimeIntegrationTest {
 
     ServiceArtifactId artifactId = ServiceArtifactId.newJavaId("com.acme/foo-service", "1.0.0");
     LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-        .newInstance(artifactId, TestServiceModule::new);
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
     ServiceInstanceSpec instanceSpec = ServiceInstanceSpec.newInstance(TEST_NAME,
         TEST_ID, artifactId);
     when(serviceLoader.findService(artifactId))
@@ -500,6 +502,119 @@ class ServiceRuntimeIntegrationTest {
     verify(serviceLoader).unloadAll();
   }
 
+  @Test
+  void migrateServiceWithoutScripts() {
+    String baseVersion = "0.0.1";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
+        .newInstance(artifactId, TestServiceModule::new, emptyList());
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.of(serviceDefinition));
+
+    Optional<MigrationScript> script = serviceRuntime.migrate(artifactId, baseVersion);
+
+    assertThat(script).isEmpty();
+  }
+
+  @Test
+  void migrateService() {
+    String baseVersion = "0.0.1";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    MigrationScript migrationScript = createScript(targetVersion);
+    LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
+        .newInstance(artifactId, TestServiceModule::new, List.of(() -> migrationScript));
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.of(serviceDefinition));
+
+    Optional<MigrationScript> script = serviceRuntime.migrate(artifactId, baseVersion);
+
+    assertThat(script).hasValue(migrationScript);
+  }
+
+  @Test
+  void migrateServiceWrongArtifact() {
+    String baseVersion = "0.0.1";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.empty());
+
+    assertThrows(IllegalArgumentException.class,
+        () -> serviceRuntime.migrate(artifactId, baseVersion));
+  }
+
+  @Test
+  void migrateServiceIncompatibleScriptsVersion() {
+    String scriptsMaxTargetVersion = "0.0.5";
+    String baseVersion = "0.1.0";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    MigrationScript migrationScript = createScript(scriptsMaxTargetVersion);
+    LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
+        .newInstance(artifactId, TestServiceModule::new, List.of(() -> migrationScript));
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.of(serviceDefinition));
+
+    assertThrows(IllegalStateException.class,
+        () -> serviceRuntime.migrate(artifactId, baseVersion));
+  }
+
+  @Test
+  void migrateServiceMinDataVersionScriptViolation() {
+    String scriptsMinSupportedVersion = "0.1.0";
+    String baseVersion = "0.5.0";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    MigrationScript migrationScript = createScriptWithMinVersion(targetVersion,
+        scriptsMinSupportedVersion);
+    LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
+        .newInstance(artifactId, TestServiceModule::new, List.of(() -> migrationScript));
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.of(serviceDefinition));
+
+    assertThrows(IllegalStateException.class,
+        () -> serviceRuntime.migrate(artifactId, baseVersion));
+  }
+
+  @Test
+  void migrateServiceDuplicateScripts() {
+    String baseVersion = "0.5.0";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    MigrationScript migrationScript1 = createScript(targetVersion);
+    MigrationScript migrationScript2 = createScript(targetVersion);
+    LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
+        .newInstance(artifactId, TestServiceModule::new,
+            List.of(() -> migrationScript1, () -> migrationScript2));
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.of(serviceDefinition));
+
+    var exception = assertThrows(IllegalStateException.class,
+        () -> serviceRuntime.migrate(artifactId, baseVersion));
+
+    assertThat(exception)
+        .hasMessageContaining(targetVersion)
+        .hasMessageContaining("duplications found: 2 scripts");
+  }
+
+  @Test
+  void migrateServiceWithMinDataVersionScript() {
+    String baseVersion = "0.5.0";
+    String targetVersion = "1.0.0";
+    ServiceArtifactId artifactId = ServiceArtifactId
+        .newJavaId("com.acme/foo-service", targetVersion);
+    MigrationScript migrationScript = createScriptWithMinVersion(targetVersion, baseVersion);
+    LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
+        .newInstance(artifactId, TestServiceModule::new, List.of(() -> migrationScript));
+    when(serviceLoader.findService(artifactId)).thenReturn(Optional.of(serviceDefinition));
+
+    Optional<MigrationScript> actualScript = serviceRuntime.migrate(artifactId, baseVersion);
+
+    assertThat(actualScript).hasValue(migrationScript);
+  }
+
   @Nested
   class WithSingleService {
     final ServiceArtifactId ARTIFACT_ID = ServiceArtifactId
@@ -520,7 +635,7 @@ class ServiceRuntimeIntegrationTest {
       when(serviceWrapper.getName()).thenReturn(TEST_NAME);
       // Setup the loader
       LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-          .newInstance(ARTIFACT_ID, TestServiceModule::new);
+          .newInstance(ARTIFACT_ID, TestServiceModule::new, emptyList());
       when(serviceLoader.findService(ARTIFACT_ID))
           .thenReturn(Optional.of(serviceDefinition));
       // Setup the factory
@@ -690,7 +805,7 @@ class ServiceRuntimeIntegrationTest {
       serviceRuntime.initialize(mock(NodeProxy.class));
 
       LoadedServiceDefinition serviceDefinition = LoadedServiceDefinition
-          .newInstance(ARTIFACT_ID, TestServiceModule::new);
+          .newInstance(ARTIFACT_ID, TestServiceModule::new, emptyList());
       when(serviceLoader.findService(ARTIFACT_ID))
           .thenReturn(Optional.of(serviceDefinition));
       // Setup the factory
@@ -752,5 +867,47 @@ class ServiceRuntimeIntegrationTest {
         .serviceId(expectedId)
         .blockchainData(expectedData)
         .build();
+  }
+
+  private static MigrationScript createScript(String version) {
+    return new AbstractMigrationScript() {
+      @Override
+      public String name() {
+        return "script-for-" + targetVersion();
+      }
+
+      @Override
+      public String targetVersion() {
+        return version;
+      }
+
+      @Override
+      public void execute(ExecutionContext context) {
+      }
+    };
+  }
+
+  private static MigrationScript createScriptWithMinVersion(String targetVersion,
+      String minVersion) {
+    return new AbstractMigrationScript() {
+      @Override
+      public String name() {
+        return "script-for-" + targetVersion();
+      }
+
+      @Override
+      public String targetVersion() {
+        return targetVersion;
+      }
+
+      @Override
+      public Optional<String> minSupportedVersion() {
+        return Optional.of(minVersion);
+      }
+
+      @Override
+      public void execute(ExecutionContext context) {
+      }
+    };
   }
 }